Fix #4 push does not preserve original timestap unlike equiv adb push from command line
Accroding to
https://android.googlesource.com/platform/system/core/+/master/adb/daemon/file_sync_service.cpp#286
adb daemon ( after sdk 27 ) always sets the atime and mtime to the `last modified time`.
So the only one thing I need to change is that
use the `os.stat` to the modified time of the source file then pass to the adb daemon as the last modified time.